refactor(app_runtime): spawn_wizard_shell_window を wizard.rs に追加 (SPEC-2077 Phase F3a)
- crates/gwt/src/app_runtime/wizard.rs に spawn_wizard_shell_window (47 行) を追加
- mod.rs から該当 method 削除
- 残: F3b (spawn_wizard_shell_window_async ~315 行) のみ

Verification:
- cargo test -p gwt-core -p gwt: 全 pass (既知 flake 1 件、 個別 pass)
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
