+++ crates/gwt/src/app_runtime/wizard.rs
@@ -23,11 +23,14 @@ use uuid::Uuid;
 
 use crate::UserEvent;
 
+use crate::ShellLaunchConfig;
+
 use super::{
-    branch_worktree_path, detect_wizard_docker_context_and_status, knowledge_error_event,
-    knowledge_kind_for_preset, list_branch_entries_with_active_sessions, normalize_branch_name,
-    preferred_issue_launch_branch, synthetic_branch_entry, AppRuntime, BackendEvent,
-    IssueLaunchWizardPrepared, LaunchWizardSession, OutboundEvent, WindowPreset,
+    branch_worktree_path, combined_window_id, detect_wizard_docker_context_and_status,
+    knowledge_error_event, knowledge_kind_for_preset, list_branch_entries_with_active_sessions,
+    normalize_branch_name, preferred_issue_launch_branch, synthetic_branch_entry, AppRuntime,
+    BackendEvent, IssueLaunchWizardPrepared, LaunchWizardSession, OutboundEvent, WindowPreset,
+    WindowProcessStatus,
 };
 
 impl AppRuntime {
@@ -418,6 +421,53 @@ impl AppRuntime {
                 vec![self.launch_wizard_state_outbound()]
             }
         }
+    }
+
+    pub(crate) fn spawn_wizard_shell_window(
+        &mut self,
+        tab_id: &str,
+        config: ShellLaunchConfig,
+        bounds: WindowGeometry,
+    ) -> Result<Vec<OutboundEvent>, String> {
+        let tab = self
+            .tab_mut(tab_id)
+            .ok_or_else(|| "Project tab not found".to_string())?;
+        let project_root = tab.project_root.display().to_string();
+        let title = format!(
+            "{} · {}",
+            config.display_name,
+            config.branch.as_ref().unwrap_or(&"workspace".to_string())
+        );
+        let window = tab
+            .workspace
+            .add_window_with_title(WindowPreset::Shell, title, false, bounds);
+        self.register_window(tab_id, &window.id);
+        let window_id = combined_window_id(tab_id, &window.id);
+
+        self.window_pty_statuses
+            .insert(window_id.clone(), WindowProcessStatus::Running);
+        self.window_hook_states.remove(&window_id);
+
+        let mut events = vec![self.workspace_state_broadcast()];
+        events.extend(Self::status_events(
+            window_id.clone(),
+            WindowProcessStatus::Running,
+            Some("Launching...".to_string()),
+        ));
+
+        let proxy = self.proxy.clone();
+        let profile_config_path = self.profile_config_path()?;
+        thread::spawn(move || {
+            Self::spawn_wizard_shell_window_async(
+                proxy,
+                project_root,
+                window_id,
+                config,
+                profile_config_path,
+            )
+        });
+
+        Ok(events)
     }
 
     pub(super) fn refresh_open_launch_wizard_from_cache(&mut self) {
